requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
lime>=0.2.0.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
grad-cam>=1.5.0
autoattack>=0.1
torchattacks>=3.4.0
scipy>=1.11.0
Pillow>=10.0.0